Text files
These are the text files to train GPT-2 on. It is human-written advice we collected in an advice-giving experiment. To generate both honest and dishonest advice, we trained the algorithm separately on (a) human-written advice encouraging first movers to be honest, and (b) human-written advice encouraging first movers to be dishonest.